refactor(ape-agent): remove dual chat backend, troop is sole target
- Delete chat-api.ts (ChatApi) and its test (tested removed behavior)
- Move shared types (PostedMessage, HistoryMessage, ContactView,
  ChatBackend) into troop-chat-api.ts
- Remove loadBridgeEnvFile (Mac ~/Library env loader)
- Remove OPENAPE_BRIDGE_TARGET / 'chat' | 'troop' union from BridgeConfig
- Bridge always uses TroopChatApi; DEFAULT_ENDPOINT → troop.openape.ai
- Update thread-session.ts + cron-runner.ts imports to troop-chat-api
